+++ db/migrate/20170420161008_add_received_not_processed_to_pledges.rb
@@ -1,0 +1,5 @@
+class AddReceivedNotProcessedToPledges < ActiveRecord::Migration
+  def change
+    add_column :pledges, :received_not_processed, :boolean
+  end
+end